requirements.txt
requests
azure-identity
IPython
azure-ai-agentserver-agentframework==1.0.0b7
agent-framework
azure-monitor-opentelemetry==1.8.1
python-dotenv
openai
azure-ai-projects>=2.0.0a20250915020
opentelemetry-instrumentation-openai-v2

agent-framework-azure-ai
